refactor: 코드 리뷰 참고하여 코드 품질 재고
