+++ MoyeoRun/Scenes/HomeTab/ViewControllers/HomeTabContainers/HomeTabNewMissionContainer.swift
@@ -1,0 +1,48 @@
+//
+//  HomeTabNewMissionContainer.swift
+//  MoyeoRun
+//
+//  Created by 김상현 on 2022/03/29.
+//
+
+import UIKit
+
+class HomeTabNewMissionContainer: UIViewController {}
+
+extension HomeTabNewMissionContainer: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 4
+    }
+
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cellIdentifier = "HomeTabNewMissionCell"
+
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellIdentifier, for: indexPath)
+            as? HomeTabNewMissionCell
+        else {
+            return UICollectionViewCell()
+        }
+        cell.setUI(forIndex: indexPath.item)
+
+        return cell
+    }
+}
+
+extension HomeTabNewMissionContainer: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let width = collectionView.bounds.width * 0.49
+        let height = collectionView.bounds.height * 0.49
+
+        return CGSize(width: width, height: height)
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+        let width = collectionView.bounds.width * 0.49
+        let totalWidth = width * 2
+        let totalSpacingWidth: CGFloat = 10
+        let leftInset = (collectionView.frame.width - CGFloat(totalWidth + totalSpacingWidth)) / 2
+        let rightInset = leftInset
+
+        return UIEdgeInsets(top: 0, left: leftInset, bottom: 0, right: rightInset)
+    }
+}
