+++ MoyeoRun/Scenes/HomeTab/ViewControllers/HomeTabContainers/HomeTabNowPopularContainer.swift
@@ -1,0 +1,42 @@
+//
+//  HomeTabNowPopularContainer.swift
+//  MoyeoRun
+//
+//  Created by 김상현 on 2022/03/29.
+//
+
+import UIKit
+
+class HomeTabNowPopularContainer: UIViewController {}
+
+extension HomeTabNowPopularContainer: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 3
+    }
+
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cellIdentifier = "HomeTabNowPopularCell"
+
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellIdentifier, for: indexPath)
+            as? HomeTabNowPopularCell
+        else {
+            return UICollectionViewCell()
+        }
+        cell.setUI(forIndex: indexPath.item)
+
+        return cell
+    }
+}
+
+extension HomeTabNowPopularContainer: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let width = collectionView.bounds.width * 0.8
+        let height = collectionView.bounds.height * 0.95
+
+        return CGSize(width: width, height: height)
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+        return UIEdgeInsets(top: 0, left: 15, bottom: 0, right: 15)
+    }
+}
