+++ MoyeoRun/Scenes/TabBar/ViewControllers/Extensions.swift
@@ -9,7 +9,7 @@ import Foundation
 import UIKit
 
 extension CALayer {
-    // Sketch 스타일의 그림자를 생성하는 유틸리티 함수
+    /// Sketch 스타일의 그림자를 생성하는 유틸리티 함수
     func applyShadow(
         color: UIColor = .black,
         alpha: Float = 0.5,
@@ -24,23 +24,23 @@ extension CALayer {
     }
 }
 
-// label에 사람 모양에다 파티원 넣는 경우에 쓰는거. "참가자.png" 참고하면 이해가 빠름.
+/// label에 사람 모양에다 파티원 넣는 경우에 쓰는거. "참가자.png" 참고하면 이해가 빠름.
 extension UILabel {
-    func setNumberOfPartyLabel(num: String){
+    func setNumberOfPartyLabel(num: String) {
         let attributedString = NSMutableAttributedString(string: "")
         let imageAttachment = NSTextAttachment()
         imageAttachment.image = UIImage(named: "참가자.png")
         imageAttachment.bounds = CGRect(x: 0, y: 0, width: 13, height: 13)
-        
+
         attributedString.append(NSAttributedString(attachment: imageAttachment))
         attributedString.append(NSAttributedString(string: num))
-        
+
         self.attributedText = attributedString
         self.sizeToFit()
     }
 }
 
-// 탭바 그림자.
+/// 탭바 그림자.
 extension UITabBar {
     // 기본 그림자 스타일을 초기화해야 커스텀 스타일을 적용할 수 있다.
     static func clearShadow() {
@@ -50,35 +50,33 @@ extension UITabBar {
     }
 }
 
-// 버튼 cornerradius를 우측 메뉴에서 조정할 수 있게해줌.
+/// 버튼 cornerradius를 우측 메뉴에서 조정할 수 있게해줌.
 @IBDesignable extension UIButton {
-
     @IBInspectable var borderWidth: CGFloat {
-        set {
-            layer.borderWidth = newValue
-        }
         get {
             return layer.borderWidth
+        }
+        set {
+            layer.borderWidth = newValue
         }
     }
 
     @IBInspectable var cornerRadius: CGFloat {
-        set {
-            layer.cornerRadius = newValue
-        }
         get {
             return layer.cornerRadius
+        }
+        set {
+            layer.cornerRadius = newValue
         }
     }
 
     @IBInspectable var borderColor: UIColor? {
-        set {
-            guard let uiColor = newValue else { return }
-            layer.borderColor = uiColor.cgColor
-        }
         get {
             guard let color = layer.borderColor else { return nil }
             return UIColor(cgColor: color)
+        }
+        set {
+            layer.borderColor = newValue?.cgColor
         }
     }
 }
