+++ MoyeoRun/Scenes/TabBar/ViewControllers/TabBarViewController.swift
@@ -6,7 +6,6 @@
 //
 
 
-import Foundation
 import UIKit
 
 class TabBarViewController: UITabBarController {
@@ -14,9 +13,11 @@ class TabBarViewController: UITabBarController {
         super.viewDidLoad()
         setUI()
     }
+
     func setUI() {
         setTabBarShadow()
     }
+
     func setTabBarShadow() {
         UITabBar.clearShadow()
         tabBar.layer.applyShadow(color: .gray, alpha: 0.3, x: 0, y: 0, blur: 12)
